+++ laxar-react-adapter.js
@@ -1,64 +1,72 @@
 /**
- * Copyright 2015 aixigo AG
+ * Copyright 2016 aixigo AG
  * Released under the MIT license.
  * http://laxarjs.org/license
  */
-
 import * as ReactDom from 'react-dom';
 
 export const technology = 'react';
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+// eslint-disable-next-line valid-jsdoc
 /**
  * Implements the LaxarJS adapter API:
  * https://github.com/LaxarJS/laxar/blob/master/docs/manuals/adapters.md
  */
-export function bootstrap( modules, laxarServices ) {
+export function bootstrap( modules ) {
 
    const widgetModules = {};
 
    modules
       .map( _ => _.default || _ )
-      .filter( function ( module ) { return module.name; } )
-      .forEach( function( module ) {
+      .filter( _ => !!_.name )
+      .forEach( module => {
          widgetModules[ module.name ] = module;
       } );
 
    return {
-      technology: technology,
-      create: create,
-      applyViewChanges: function() {}
+      technology,
+      create,
+      applyViewChanges() {}
    };
 
    ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-   function create( environment, services ) {
+   function create( environment ) {
 
-      const widgetName = environment.specification.name;
-      const context = environment.context;
       let isAttached = true;
       let controller = null;
       return {
-         createController: createController,
-         domAttachTo: domAttachTo,
-         domDetach: domDetach,
-         destroy: function() {}
+         createController,
+         domAttachTo,
+         domDetach,
+         destroy() {}
       };
 
-      /////////////////////////////////////////////////////////////////////////////////////////////////////
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       function createController( config ) {
-         const widgetModule = widgetModules[ widgetName ];
-         const injector = createInjector();
-         const injections = ( widgetModule.injections || [] ).map( function( injection ) {
-            return injector.get( injection );
+         const module = widgetModules[ environment.specification.name ];
+         const availableInjections = {
+            ...environment.services,
+            axReactRender( componentInstance ) {
+               if( isAttached ) {
+                  ReactDom.render( componentInstance, environment.anchorElement );
+               }
+            }
+         };
+         const injections = ( module.injections || [] ).map( injection => {
+            if( !( injection in availableInjections ) ) {
+               throw new Error( `Trying to inject unknown service "${injection}".` );
+            }
+            return availableInjections[ injection ];
          } );
-         config.onBeforeControllerCreation( environment, injector.get() );
-         controller = widgetModule.create.apply( widgetModule, injections );
+         config.onBeforeControllerCreation( environment, Object.freeze( injections ) );
+         controller = module.create( ...injections );
       }
 
-      /////////////////////////////////////////////////////////////////////////////////////////////////////
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       function domAttachTo( areaElement ) {
          isAttached = true;
@@ -66,7 +74,7 @@ export function bootstrap( modules, laxarServices ) {
          controller.onDomAvailable();
       }
 
-      /////////////////////////////////////////////////////////////////////////////////////////////////////
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       function domDetach() {
          isAttached = false;
@@ -74,46 +82,6 @@ export function bootstrap( modules, laxarServices ) {
          if( parent ) {
             parent.removeChild( environment.anchorElement );
          }
-      }
-
-      /////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      function createInjector() {
-         const map = {
-            axLog: laxarServices.log,
-            axConfiguration: laxarServices.configuration,
-            axContext: context,
-            axEventBus: context.eventBus,
-            axFeatures: context.features || {},
-            axReactRender: function( componentInstance ) {
-               if( isAttached ) {
-                  ReactDom.render(
-                     componentInstance,
-                     environment.anchorElement
-                  );
-               }
-            }
-         };
-
-         //////////////////////////////////////////////////////////////////////////////////////////////////
-
-         return {
-            get: function( name ) {
-               if( arguments.length === 0 ) {
-                  return map;
-               }
-
-               if( name in map ) {
-                  return map[ name ];
-               }
-
-               if( name in services ) {
-                  return services[ name ];
-               }
-
-               throw new Error( 'laxar-react-adapter: Unknown dependency: "' + name + '"' );
-            }
-         };
       }
 
    }