+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#5](https://github.com/LaxarJS/laxar-react-adapter/issues/5): services: adopted to new widget services
+   + **BREAKING CHANGE:** see ticket for details
 - [#7](https://github.com/LaxarJS/laxar-react-adapter/issues/6): added dist-target for UMD-library
 - [#6](https://github.com/LaxarJS/laxar-react-adapter/issues/6): made react adapter module stateless (LaxarJS v2)
    + **BREAKING CHANGE:** see ticket for details